[control-plane-manager] Move update-observer to control-plane-manager (#19394)
