+++ modules/040-control-plane-manager/images/control-plane-manager/src/internal/controllers/update-observer/cluster/controlplane.go
@@ -23,7 +23,7 @@ import (
 
 	corev1 "k8s.io/api/core/v1"
 
-	"update-observer/pkg/version"
+	"control-plane-manager/internal/controllers/update-observer/pkg/version"
 )
 
 type ControlPlaneState struct {
