+++ modules/040-control-plane-manager/images/control-plane-manager/src/internal/controllers/update-observer/cluster/env.go
@@ -21,7 +21,7 @@ import (
 	"os"
 	"strings"
 
-	"update-observer/pkg/version"
+	"control-plane-manager/internal/controllers/update-observer/pkg/version"
 )
 
 type VersionSettings struct {
