+++ modules/040-control-plane-manager/images/control-plane-manager/src/internal/controllers/update-observer/cluster/master_node.go
@@ -22,8 +22,8 @@ import (
 
 	corev1 "k8s.io/api/core/v1"
 
-	podstatus "update-observer/pkg/pod-status"
-	"update-observer/pkg/version"
+	podstatus "control-plane-manager/internal/controllers/update-observer/pkg/pod-status"
+	"control-plane-manager/internal/controllers/update-observer/pkg/version"
 )
 
 type MasterNode struct {
