+++ modules/040-control-plane-manager/images/control-plane-manager/src/internal/controllers/update-observer/controller/controller.go
@@ -20,10 +20,10 @@ import (
 	"context"
 	"time"
 
+	"github.com/deckhouse/deckhouse/pkg/log"
 	"golang.org/x/time/rate"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/client-go/util/workqueue"
-	"k8s.io/klog/v2"
 	"k8s.io/utils/ptr"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/builder"
@@ -35,10 +35,10 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/predicate"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
-	"update-observer/cluster"
-	"update-observer/common"
-	v1 "update-observer/pkg/v1"
-	"update-observer/pkg/version"
+	"control-plane-manager/internal/controllers/update-observer/cluster"
+	"control-plane-manager/internal/controllers/update-observer/common"
+	v1 "control-plane-manager/internal/controllers/update-observer/pkg/v1"
+	"control-plane-manager/internal/controllers/update-observer/pkg/version"
 )
 
 const (
@@ -55,6 +55,8 @@ const (
 	ReconcileTriggerDowngradeK8s ReconcileTrigger = "downgradeK8s"
 	ReconcileTriggerIdle         ReconcileTrigger = "idle"
 )
+
+var logger = log.NewLogger().Named("update-observer")
 
 type reconciler struct {
 	client client.Client
@@ -158,17 +160,15 @@ func getNodeGroupPredicate() predicate.Predicate {
 }
 
 func (r *reconciler) Reconcile(ctx context.Context, req reconcile.Request) (reconcile.Result, error) {
-	klog.Info("Reconcile started")
-
 	configMap, err := r.getConfigMap(ctx)
 	if err != nil {
-		klog.Error("Failed to get configMap", err)
+		logger.Error("Failed to get configMap", "namespace", common.KubeSystemNamespace, "name", common.ConfigMapName, log.Err(err))
 		return reconcile.Result{RequeueAfter: requeueInterval}, nil
 	}
 
 	clusterCfg, err := r.getClusterConfiguration(ctx)
 	if err != nil {
-		klog.Info("Error occurred while getting cluster configuration", err)
+		logger.Error("Failed to get cluster configuration from secret", "namespace", common.KubeSystemNamespace, "name", common.SecretName, log.Err(err))
 		return reconcile.Result{RequeueAfter: requeueInterval}, nil
 	}
 
@@ -176,22 +176,23 @@ func (r *reconciler) Reconcile(ctx context.Context, req reconcile.Request) (reco
 
 	clusterState, err := r.getClusterState(ctx, clusterCfg, configMap.Labels, reconcileTrigger == ReconcileTriggerDowngradeK8s)
 	if err != nil {
-		klog.Error("Error encountered while getting cluster state", err)
+		logger.Error("Failed to get cluster state", log.Err(err))
 		return reconcile.Result{RequeueAfter: requeueInterval}, nil
 	}
 
 	configMap, err = fillConfigMap(configMap, clusterState, reconcileTrigger)
 	if err != nil {
-		klog.Error("Failed to fill configMap", err)
+		logger.Error("Failed to fill configMap", log.Err(err))
 		return reconcile.Result{RequeueAfter: requeueInterval}, nil
 	}
 
 	if err = r.touchConfigMap(ctx, configMap); err != nil {
-		klog.Error("Failed to touch configMap", err)
+		logger.Error("Failed to write configMap", "namespace", common.KubeSystemNamespace, "name", common.ConfigMapName, log.Err(err))
 		return reconcile.Result{RequeueAfter: requeueInterval}, nil
 	}
 
 	if clusterState.Status.Phase != cluster.ClusterUpToDate {
+		logger.Info("Cluster not up-to-date, requeuing", "after", requeueInterval)
 		return reconcile.Result{RequeueAfter: requeueInterval}, nil
 	}
