+++ modules/040-control-plane-manager/images/control-plane-manager/src/internal/controllers/update-observer/controller/testdata/cases/pods-failed.yaml
