--- modules/040-control-plane-manager/images/update-observer/src/main.go
@@ -1,51 +1,0 @@
-/*
-Copyright 2026 Flant JSC
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package main
-
-import (
-	"context"
-	"os"
-	"os/signal"
-	"syscall"
-
-	"k8s.io/klog/v2"
-
-	"update-observer/common"
-)
-
-func main() {
-	ctx, cancel := context.WithCancel(context.Background())
-
-	manager, err := NewManager(ctx) // TODO pprof flag, removed bool because of linting
-	if err != nil {
-		klog.Fatalf("Failed to create a manager: %v", err)
-	}
-
-	if err = manager.Start(ctx); err != nil {
-		klog.Fatalf("Failed to start the manager: %v", err)
-	}
-
-	sigs := make(chan os.Signal, 1)
-	signal.Notify(sigs, syscall.SIGINT, syscall.SIGTERM)
-
-	for range sigs {
-		klog.Info("Shutdown signal received")
-		cancel()
-		klog.Infof("Bye from %s", common.ControllerName)
-		break
-	}
-}
