--- modules/040-control-plane-manager/images/update-observer/src/manager.go
@@ -1,163 +1,0 @@
-/*
-Copyright 2026 Flant JSC
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package main
-
-import (
-	"context"
-	"fmt"
-	"time"
-
-	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
-	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
-	"k8s.io/klog/v2"
-	"k8s.io/klog/v2/textlogger"
-	"k8s.io/utils/ptr"
-	controllerruntime "sigs.k8s.io/controller-runtime"
-	"sigs.k8s.io/controller-runtime/pkg/cache"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/manager"
-	metricsserver "sigs.k8s.io/controller-runtime/pkg/metrics/server"
-
-	"update-observer/common"
-	"update-observer/controller"
-	v1 "update-observer/pkg/v1"
-)
-
-var (
-	scheme = runtime.NewScheme()
-)
-
-func init() {
-	utilruntime.Must(clientgoscheme.AddToScheme(scheme))
-	utilruntime.Must(v1.AddToScheme(scheme))
-}
-
-type Manager struct {
-	runtimeManager manager.Manager
-}
-
-func NewManager(ctx context.Context) (*Manager, error) {
-	cfg := controllerruntime.GetConfigOrDie()
-	controllerruntime.SetLogger(textlogger.NewLogger(textlogger.NewConfig()))
-
-	runtimeManager, err := controllerruntime.NewManager(cfg, controllerruntime.Options{
-		Scheme:           scheme,
-		LeaderElection:   true,
-		LeaderElectionID: common.ControllerName,
-		BaseContext: func() context.Context {
-			return ctx
-		},
-		Metrics: metricsserver.Options{
-			BindAddress: "0",
-		},
-		HealthProbeBindAddress:  "0",
-		PprofBindAddress:        "0",
-		GracefulShutdownTimeout: ptr.To(10 * time.Second),
-		Cache: cache.Options{
-			ReaderFailOnMissingInformer: false,
-			DefaultTransform:            cache.TransformStripManagedFields(),
-			ByObject: map[client.Object]cache.ByObject{
-				&corev1.Secret{}: {
-					Namespaces: map[string]cache.Config{
-						common.KubeSystemNamespace: {},
-					},
-				},
-				&corev1.Node{}: {
-					Transform: func(in any) (any, error) {
-						node, ok := in.(*corev1.Node)
-						if !ok {
-							return in, nil
-						}
-
-						stripped := &corev1.Node{}
-						stripped.Name = node.Name
-						stripped.ResourceVersion = node.ResourceVersion
-						stripped.UID = node.UID
-						stripped.Status = corev1.NodeStatus{
-							NodeInfo: corev1.NodeSystemInfo{
-								KubeletVersion: node.Status.NodeInfo.KubeletVersion,
-							},
-						}
-
-						return stripped, nil
-					},
-				},
-				&v1.NodeGroup{}: {
-					Transform: func(in any) (any, error) {
-						ng, ok := in.(*v1.NodeGroup)
-						if !ok {
-							return in, nil
-						}
-
-						stripped := &v1.NodeGroup{}
-						stripped.Name = ng.Name
-						stripped.ResourceVersion = ng.ResourceVersion
-						stripped.UID = ng.UID
-						stripped.Status.Ready = ng.Status.Ready
-
-						return stripped, nil
-					},
-				},
-			},
-		},
-		Client: client.Options{
-			Cache: &client.CacheOptions{
-				DisableFor: []client.Object{
-					&corev1.Pod{},
-					&corev1.ConfigMap{},
-				},
-			},
-		},
-	})
-	if err != nil {
-		return nil, fmt.Errorf("create controller runtime manager: %w", err)
-	}
-
-	// if err = runtimeManager.AddHealthzCheck("healthz", healthz.Ping); err != nil {
-	// 	return nil, fmt.Errorf("add health check: %w", err)
-	// }
-
-	// if err = runtimeManager.AddReadyzCheck("readyz", healthz.Ping); err != nil {
-	// 	return nil, fmt.Errorf("add ready check: %w", err)
-	// }
-
-	if err = controller.RegisterController(runtimeManager); err != nil {
-		return nil, fmt.Errorf("add controller: %w", err)
-	}
-
-	return &Manager{
-		runtimeManager,
-	}, nil
-}
-
-func (c *Manager) Start(ctx context.Context) error {
-	go func() {
-		if err := c.runtimeManager.Start(ctx); err != nil {
-			klog.Fatalf("failed to start runtime manager: %v", err)
-		}
-	}()
-	klog.Info("Runtime manager started")
-
-	if ok := c.runtimeManager.GetCache().WaitForCacheSync(ctx); !ok {
-		return fmt.Errorf("wait for cache sync")
-	}
-	klog.Info("Cache synced")
-
-	return nil
-}
