--- modules/040-control-plane-manager/images/update-observer/werf.inc.yaml
@@ -1,39 +1,0 @@
----
-image: {{ $.ModuleName }}/{{ $.ImageName }}
-fromImage: common/distroless
-import:
-- image: {{ $.ModuleName }}/{{ $.ImageName }}-artifact
-  add: /src/update-observer
-  to: /update-observer
-  before: install
-imageSpec:
-  config:
-    entrypoint: ["/update-observer"]
----
-image: {{ $.ModuleName }}/{{ $.ImageName }}-artifact
-from: {{ index $.Images "builder/golang-alpine" }}
-final: false
-import:
-- image: {{ $.ModuleName }}/{{ $.ImageName }}-src-artifact
-  add: /src
-  to: /src
-  before: install
-mount:
-{{ include "mount points for golang builds" . }}
-secrets:
-- id: GOPROXY
-  value: {{ .GOPROXY }}
-shell:
-  install:
-    - cd /src
-    - GOPROXY=$(cat /run/secrets/GOPROXY) GOOS=linux GOARCH=amd64 CGO_ENABLED=0 go build -ldflags="-s -w" -o update-observer
----
-image: {{ $.ModuleName }}/{{ $.ImageName }}-src-artifact
-fromImage: common/src-artifact
-final: false
-git:
-- add: /{{ $.ModulePath }}modules/040-{{ $.ModuleName }}/images/{{ $.ImageName }}/src
-  to: /src
-  stageDependencies:
-    install:
-    - '**/*'
