+++ modules/040-control-plane-manager/templates/daemonset.yaml
@@ -8,10 +8,6 @@ cpu: 10m
 memory: 10Mi
 {{- end }}
 
-{{- define "update_observer_resources" }}
-cpu: 10m
-memory: 10Mi
-{{- end }}
 
 {{- $profiles := list "main" }}
 {{- if .Values.controlPlaneManager.internal.hasEtcdArbiterNode }}
@@ -226,12 +222,6 @@ spec:
         cpu: 10m
         memory: 10Mi
     {{- end }}
-    - containerName: update-observer
-      minAllowed:
-        {{- include "update_observer_resources" $ | nindent 8 }}
-      maxAllowed:
-        cpu: 50m
-        memory: 80Mi
 {{- end }}
 ---
 apiVersion: apps/v1
@@ -305,6 +295,8 @@ spec:
           value: {{ $.Values.global.clusterConfiguration.kubernetesVersion | quote }}
         - name: ALLOWED_KUBERNETES_VERSIONS
           value: {{ $allowedKubernetesVersions | sortAlpha | join  "," | quote }}
+        - name: AUTOMATIC_KUBERNETES_VERSION
+          value: {{ $automaticVersion | quote }}
         - name: NODE_NAME
           valueFrom:
             fieldRef:
@@ -340,27 +332,6 @@ spec:
           requests:
             {{- include "image_holder_resources" $ | nindent 12 }}
 {{- end }}
-      - name: update-observer
-        {{ include "helm_lib_module_container_security_context_pss_restricted_flexible" dict  | nindent 8 }}
-        image: {{ include "helm_lib_module_image" (list $ "updateObserver") }}
-        command:
-        - /update-observer
-        resources:
-          requests:
-            {{- include "helm_lib_module_ephemeral_storage_only_logs" $ | nindent 12 }}
-  {{- if not ( $.Values.global.enabledModules | has "vertical-pod-autoscaler") }}
-            {{- include "update_observer_resources" $ | nindent 12 }}
-  {{- end }}
-        env:
-        - name: LEADER_ELECTION_NAMESPACE
-          valueFrom:
-            fieldRef:
-              fieldPath: metadata.namespace
-        - name: ALLOWED_KUBERNETES_VERSIONS
-          value: {{ $allowedKubernetesVersions | sortAlpha | join  "," | quote }}
-        - name: AUTOMATIC_KUBERNETES_VERSION
-          value: {{ $automaticVersion | quote }}
-        {{- include "helm_lib_envs_for_proxy" $ | nindent 8 }}
       hostNetwork: true
       dnsPolicy: ClusterFirstWithHostNet
       priorityClassName: system-cluster-critical
