+++ modules/040-control-plane-manager/templates/rbac-for-us.yaml
@@ -25,7 +25,7 @@ rules:
   verbs: ["get", "watch", "list"]
 - apiGroups: [""]
   resources: ["configmaps"]
-  verbs: ["get", "create", "update"]
+  verbs: ["get", "list", "watch", "create", "update"]
 - apiGroups: [""]
   resources: ["events"]
   verbs: ["create"]
